Index

OscillatoryIntegralsODE.BesselJIntegral

OscillatoryIntegralsODE.HarmonicIntegral

OscillatoryIntegralsODE.SphericalBesselJIntegral

OscillatoryIntegralsODE.levintegrate

OscillatoryIntegralsODE.BesselJIntegral — Type

BesselJIntegral{T}(ν, r)

This constructs a Bessel integration problem which solves

\[I = \int_a^b f(x) J_{\nu}(r x)\, dx\]

and produces output of type {T}.

OscillatoryIntegralsODE.HarmonicIntegral — Method

HarmonicIntegral{T}(ω)

This constructs a spherical Bessel integration problem which solves

\[I = \int_a^b f(x) e^{i \omega x} \, dx\]

and produces output of type {T}.

OscillatoryIntegralsODE.SphericalBesselJIntegral — Type

SphericalBesselJIntegral{T}(ν, r)

This constructs a spherical Bessel integration problem which solves

\[I = \int_a^b f(x) j_{\nu}(r x)\, dx\]

and produces output of type {T}.

OscillatoryIntegralsODE.levintegrate — Function

levintegrate(oi, f, a, b, alg=Vern9(); ode_kwargs...)

This computes $\int_a^b f(x) w(x) dx$ for $0 < a < b$, where $f(x)$ is a smooth, univariate function, and $w(x)$ is a highly-oscillatory function.

Arguments:

oi::OscillatoryIntegral: parameters of the highly-oscillatory integral

f: the function to be integrated

a: lower integration bound, must be greater than zero

b: upper integration bound, must be greater than a

alg: algorithm used in the ODE solve, you probably want something high order

Keywords

Any keyword arguments are passed to the solve call of the Levin ODE

Returns:

the result of the integral

Examples

julia> f(x) = exp(-x^2/16)
       bi = BesselJIntegral{Float64}(100., 100.)  # nu, r
       levintegrate(bi, f, 1.0, 5.0, Vern9(); abstol=1e-6, reltol=1e-6)

0.006311599451652101